+++ components/txn_types/src/write.rs
@@ -1,9 +1,9 @@
 // Copyright 2016 TiKV Project Authors. Licensed under Apache-2.0.
 
-use super::lock::LockType;
-use super::timestamp::TimeStamp;
-use super::types::{Value, SHORT_VALUE_MAX_LEN, SHORT_VALUE_PREFIX};
-use super::{Error, Result};
+use crate::lock::LockType;
+use crate::timestamp::TimeStamp;
+use crate::types::{Value, SHORT_VALUE_MAX_LEN, SHORT_VALUE_PREFIX};
+use crate::{Error, Result};
 use codec::prelude::NumberDecoder;
 use tikv_util::codec::number::{NumberEncoder, MAX_VAR_U64_LEN};
